+++ mapchete_hub/cli/main.py
@@ -1,34 +1,53 @@
 import click
+import click_spinner
+import logging
 from mapchete import Timer
-from mapchete.cli.utils import opt_debug
+from mapchete.cli import utils
 from tqdm import tqdm
+import warnings
 
 import mapchete_hub
 from mapchete_hub.api import API, job_states
 from mapchete_hub.config import host_options
 from mapchete_hub.exceptions import JobFailed
+from mapchete_hub import log
 
 # https://github.com/tqdm/tqdm/issues/481
 tqdm.monitor_interval = 0
 
 
-@click.version_option(version=mapchete_hub.__version__, message='%(version)s')
+def _set_debug_log_level(ctx, param, debug):
+    if debug:
+        log.set_log_level(logging.DEBUG)
+    return debug
+
+
+opt_debug = click.option(
+    "--debug", "-d",
+    is_flag=True,
+    callback=_set_debug_log_level,
+    help="Print debug log output."
+)
+
+
+@click.version_option(version=mapchete_hub.__version__, message="%(version)s")
 @click.group(help="Process control on Mapchete Hub.")
 @click.option(
-    '--host', '-h',
+    "--host", "-h",
     type=click.STRING,
     nargs=1,
-    default='%s:%s' % (host_options["host_ip"], host_options["port"]),
+    default="%s:%s" % (host_options["host_ip"], host_options["port"]),
     help="Address and port of mhub endpoint (default: %s:%s)." % (
         host_options["host_ip"], host_options["port"]
     )
 )
 @click.pass_context
 def mhub(ctx, **kwargs):
+    """Main command group."""
     ctx.obj = dict(**kwargs)
 
 
-@mhub.command(short_help='Show available processes.')
+@mhub.command(short_help="Show available processes.")
 @click.option(
     "--process_name", "-n", type=click.STRING, help="Print docstring of process."
 )
@@ -37,7 +56,7 @@ def mhub(ctx, **kwargs):
 )
 @click.pass_context
 def processes(ctx, process_name=None, docstrings=False):
-
+    """Show available processes."""
     def _print_process_info(process_module, docstrings=False):
         click.echo(
             click.style(
@@ -70,9 +89,10 @@ def processes(ctx, process_name=None, docstrings=False):
         click.echo("Error: %s" % e)
 
 
-@mhub.command(short_help='Show available processes.')
+@mhub.command(short_help="Show available queues and workers.")
 @click.pass_context
 def queues(ctx):
+    """Show available queues and workers."""
     try:
         res = API(host=ctx.obj["host"]).get("capabilities.json")
         if res.status_code != 200:
@@ -88,24 +108,115 @@ def queues(ctx):
         click.echo("Error: %s" % e)
 
 
-@mhub.command(short_help='Starts job.')
-@click.argument('job_id', type=click.STRING)
-@click.argument('mapchete_file', type=click.STRING)
-@click.option('--bounds', '-b', type=float, nargs=4)
-@click.option(
-    '--mode', '-m', type=click.Choice(["continue", "overwrite"]), default="overwrite"
-)
+@mhub.command(help="Execute a process.")
+@utils.arg_mapchete_files
+@utils.opt_zoom
+@utils.opt_bounds
+@utils.opt_point
+@utils.opt_wkt_geometry
+@utils.opt_tile
+@utils.opt_overwrite
+@utils.opt_verbose
 @opt_debug
+@click.option(
+    "--queue", "-q",
+    type=click.STRING,
+    default="execute_queue",
+    help="Queue the job should be added to."
+)
+@click.pass_context
+def execute(
+    ctx,
+    mapchete_files,
+    overwrite=False,
+    verbose=False,
+    queue=None,
+    **kwargs
+):
+    """Execute a process."""
+    for mapchete_file in mapchete_files:
+        try:
+            job = API(host=ctx.obj["host"]).start_job(
+                mapchete_file,
+                mode="overwrite" if overwrite else "continue",
+                mhub_worker="execute_worker",
+                mhub_queue=queue,
+                **kwargs
+            )
+            if verbose:
+                click.echo("job %s state: %s" % (job.job_id, job.state))
+                _show_progress(ctx, job.job_id)
+            else:
+                click.echo(job.job_id)
+        except Exception as e:
+            click.echo("Error: %s" % e)
+
+
+@mhub.command(help="Create index of output tiles.")
+@utils.arg_mapchete_files
+@utils.opt_zoom
+@utils.opt_bounds
+@utils.opt_point
+@utils.opt_wkt_geometry
+@utils.opt_tile
+@utils.opt_verbose
+@click.option(
+    "--queue", "-q",
+    type=click.STRING,
+    default="index_queue",
+    help="Queue the job should be added to."
+)
+@click.pass_context
+def index(
+    ctx,
+    mapchete_files,
+    verbose=False,
+    queue=None,
+    **kwargs
+):
+    """Create index of output tiles."""
+    for mapchete_file in mapchete_files:
+        try:
+            job = API(host=ctx.obj["host"]).start_job(
+                mapchete_file,
+                mhub_worker="index_worker",
+                mhub_queue=queue,
+                **kwargs
+            )
+            if verbose:
+                click.echo("job %s state: %s" % (job.job_id, job.state))
+                _show_progress(ctx, job.job_id)
+            else:
+                click.echo(job.job_id)
+        except Exception as e:
+            click.echo("Error: %s" % e)
+
+
+@mhub.command(short_help="Start job.")
+@click.argument("job_id", type=click.STRING)
+@click.argument("mapchete_file", type=click.STRING)
+@click.option("--bounds", "-b", type=float, nargs=4)
+@click.option(
+    "--mode", "-m",
+    type=click.Choice(["continue", "overwrite"]),
+    default="overwrite"
+)
+@utils.opt_debug
 @click.pass_context
 def start(ctx, job_id, mapchete_file, bounds=None, mode=None, debug=False):
+    """Start job."""
+    warnings.warn(DeprecationWarning(
+        "This command will be deprecated. Please use `mhub execute` instead."
+    ))
     try:
         click.echo(
             "job %s state: %s" % (
                 job_id,
                 API(host=ctx.obj["host"]).start_job(
-                    job_id,
                     mapchete_file,
-                    bounds, mode=mode
+                    job_id=job_id,
+                    bounds=bounds,
+                    mode=mode
                 ).state
             )
         )
@@ -113,12 +224,13 @@ def start(ctx, job_id, mapchete_file, bounds=None, mode=None, debug=False):
         click.echo("Error: %s" % e)
 
 
-@mhub.command(short_help='Shows job status.')
-@click.argument('job_id', type=click.STRING)
-@click.option('--geojson', is_flag=True)
-@click.option('--traceback', is_flag=True)
+@mhub.command(short_help="Show job status.")
+@click.argument("job_id", type=click.STRING)
+@click.option("--geojson", is_flag=True)
+@click.option("--traceback", is_flag=True)
 @click.pass_context
 def status(ctx, job_id, geojson=False, traceback=False):
+    """Show job status."""
     try:
         response = (
             API(host=ctx.obj["host"]).job(job_id, geojson=geojson)
@@ -138,23 +250,27 @@ def status(ctx, job_id, geojson=False, traceback=False):
         click.echo("Error: %s" % e)
 
 
-@mhub.command(short_help='Shows job progress.')
-@click.argument('job_id', type=click.STRING)
+@mhub.command(short_help="Show job progress.")
+@click.argument("job_id", type=click.STRING)
 @click.pass_context
 def progress(ctx, job_id):
+    """Show job progress."""
     try:
-        show_progress(ctx, job_id)
+        _show_progress(ctx, job_id)
     except Exception as e:
         click.echo("Error: %s" % e)
 
 
-@mhub.command(short_help='Shows current jobs.')
-@click.option('--geojson', is_flag=True)
+@mhub.command(short_help="Show current jobs.")
+@click.option("--geojson", is_flag=True)
 @click.option(
-    '--output_path', type=click.STRING, help="only print jobs with specific output_path"
+    "--output_path",
+    type=click.STRING,
+    help="only print jobs with specific output_path"
 )
 @click.pass_context
 def jobs(ctx, geojson=False, output_path=None):
+    """Show current jobs."""
     try:
         click.echo(
             API(host=ctx.obj["host"]).jobs(geojson=geojson, output_path=output_path)
@@ -170,10 +286,11 @@ def jobs(ctx, geojson=False, output_path=None):
         click.echo("Error: %s" % e)
 
 
-def show_progress(ctx, job_id):
+def _show_progress(ctx, job_id):
     try:
-        states = API(host=ctx.obj["host"]).job_progress(job_id)
-        i = next(states)
+        with click_spinner.spinner():
+            states = API(host=ctx.obj["host"]).job_progress(job_id)
+            i = next(states)
         if i["state"] == "SUCCESS":
             click.echo(
                 "job %s successfully finished in %s" % (